Update index.md
edit so links can be placed on the landing page of the wiki for quick access.
hopefully the format for links work with images or ill look a bit silly.

sprites will have to be added later.

i think it should work but im unsure how it will show up in window.

whoever wrote this: "but most of the time you are just trapped in a room and never able to break free or just thrown into space for no reason at all." 

NO JUST NO. DO NOT WRITE THAT LANGUAGE. IT MAKES PEOPLE LESS LIKELY TO PLAY THE GAME LIKE WDF!!!!
its like saying "if you play this, you'll be locked in a box and be able to do nothing.. not even play the game or roleplay or do anything at all. So theres no reason to start."

Whats the point of playing the game if its like that?

Feel free to edit this as needed.