Cutting the memes and opinionated material
This wiki supports the occasional joke here or there, but full on meme _sections_ on important pages such as Adminning is not approrpiate. The wiki serves as an educational tool first. Memes come at like 5th place in the hierarchy.

In addition, the wiki's intent is to stay neutral on topics that are a matter of opinion, such has to admin.  There should not be references to 'I', or 'me'. speculation, such as the sentence, "I expect more damage has been done by those types of admins than any mechanic or gameplay element of the game that people don't like" does not have a place in this wiki. Keep that kind of stuff in the community chat.

"It's a small guide with a little advice, if anyone wants to add some to it or perhaps disagree with the advice, suggest an edit!" this being a wiki, comments like this are unnecessary and redundant.

"Generally, the game works better without an admin. **That's actually a fact.** Players are people. People know how to generally act fine." this an _extremely_ controversial thing to put forward. Please do not put stuff like this in the wiki

The admin page should be primarily explaining how to use admin tools, not how to admin.